backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pyodbc
python-dotenv
python-jose[cryptography]
python-multipart